requirements.txt
Django==2.0
django-ckeditor==5.6.1
django-cleanup==2.1.0
django-crispy-forms==1.7.2
django-js-asset==1.1.0
Pillow==5.2.0
pytz==2018.5
